split benches and remove riscv dep from pipeline (#1350)
This PR

- removes the riscv dev dependency from the pipeline crate
- splits the benchmark into riscv and pil benches
- fixes the query callback in the keccak bench
