+++ pipeline/Cargo.toml
@@ -36,8 +36,6 @@ serde_cbor = "0.11.2"
 num-traits = "0.2.15"
 
 [dev-dependencies]
-powdr-riscv = { path = "../riscv" }
-
 test-log = "0.2.12"
 env_logger = "0.10.0"
 criterion = { version = "0.4", features = ["html_reports"] }
@@ -49,5 +47,5 @@ development = ["env_logger"]
 walkdir = "2.4.0"
 
 [[bench]]
-name = "executor_benchmark"
+name = "evaluator_benchmark"
 harness = false
